# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Library API
  description: |
    API REST para el sistema de gestión de biblioteca digital personal.
    
    Permite catalogar, organizar y buscar libros digitales usando búsqueda semántica
    potenciada por embeddings vectoriales.
  version: 0.1.0
  contact:
    name: Library Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Renovar tokens de autenticación
      description: |
        Renueva el par de tokens leyendo automáticamente la cookie `refresh_token`.
        No requiere body.

        Si el token de refresco es válido y no ha expirado, se emiten nuevos tokens
        y se sobreescriben las cookies existentes.

        ## Flujo típico

        1. El cliente detecta un error `401` en cualquier endpoint protegido.
        2. Llama a `POST /auth/refresh` (la cookie `refresh_token` se envía automáticamente).
        3. Si tiene éxito, reintenta la petición original con las nuevas cookies.
      operationId: refreshToken
      responses:
        '200':
          description: Tokens renovados exitosamente. Nuevas cookies establecidas.
          headers:
            Set-Cookie:
              description: |
                Nuevas cookies con los tokens renovados:
                - `access_token=<nuevo-jwt>; HttpOnly; Secure; SameSite=Strict; Max-Age=604800; Path=/`
                - `refresh_token=<nuevo-token>; HttpOnly; Secure; SameSite=Strict; Max-Age=2592000; Path=/`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
              example:
                success: true
                data:
                  message: "Token refreshed"
                error: null
        '401':
          description: Token de refresco ausente, inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                missing_token:
                  summary: Cookie refresh_token ausente
                  value:
                    success: false
                    data: null
                    error:
                      message: "Refresh token missing"
                invalid_token:
                  summary: Token inválido o expirado
                  value:
                    success: false
                    data: null
                    error:
                      message: "Refresh token inválido o expirado"
...
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Solicitar restablecimiento de contraseña
      description: |
        Inicia el flujo de recuperación de contraseña para el email indicado.

        **Siempre retorna 200** independientemente de si el email existe o no,
        para evitar la enumeración de usuarios (user enumeration attack).

        Si el email corresponde a un usuario registrado, se enviará un correo
        con un enlace de restablecimiento.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: "usuario@example.com"
      responses:
        '200':
          description: |
            Solicitud procesada. Si el email existe, se envía un correo con el enlace de restablecimiento.
            Esta respuesta es idéntica tanto si el email existe como si no.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
              example:
                success: true
                data:
                  message: "If that email is registered, a reset link has been sent."
                error: null
        '422':
          description: Error de validación del body (email inválido o ausente).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                success: false
                data: null
                error:
                  message: "Validation failed"
                  details:
                    - "email: Invalid email address"
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Restablecer contraseña
      description: |
        Valida el token de restablecimiento y actualiza la contraseña del usuario.

        El token se obtiene del enlace enviado por `POST /auth/forgot-password`.
        Una vez usado (o expirado), el token queda invalidado y no puede reutilizarse.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              token: "abc123def456..."
              newPassword: "nueva-contraseña-segura"
      responses:
        '200':
          description: Contraseña restablecida exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
              example:
                success: true
                data:
                  message: "Password reset successful"
                error: null
        '400':
          description: Token de restablecimiento inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                invalid_token:
                  summary: Token inválido
                  value:
                    success: false
                    data: null
                    error:
                      message: "Token de restablecimiento inválido"
                expired_token:
                  summary: Token expirado
                  value:
                    success: false
                    data: null
                    error:
                      message: "Token de restablecimiento expirado"
        '422':
          description: Error de validación del body (token vacío, contraseña muy corta, etc.).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                missing_token:
                  summary: Token vacío
                  value:
                    success: false
                    data: null
                    error:
                      message: "Validation failed"
                      details:
                        - "token: Token is required"
                password_too_short:
                  summary: Contraseña muy corta
                  value:
                    success: false
                    data: null
                    error:
                      message: "Validation failed"
                      details:
                        - "newPassword: Password must be at least 8 characters"
...
  /books/{id}/favorite:
    post:
      tags:
        - Favorites
      summary: Alternar favorito de un libro
      description: |
        Alterna el estado de favorito del libro indicado para el usuario autenticado.

        - Si el libro **no está** en favoritos del usuario, lo agrega y retorna `{ "favorite": true }`.
        - Si el libro **ya está** en favoritos del usuario, lo elimina y retorna `{ "favorite": false }`.

        ## Autenticación

        Requiere un JWT válido en la cookie `access_token`. Si el token está ausente o es inválido,
        se retorna `401 Unauthorized`.
      operationId: toggleFavorite
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del libro (UUID v4)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Estado de favorito actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteToggleResponse'
              examples:
                added_to_favorites:
                  summary: Libro añadido a favoritos
                  value:
                    success: true
                    data:
                      favorite: true
                    error: null
                removed_from_favorites:
                  summary: Libro eliminado de favoritos
                  value:
                    success: true
                    data:
                      favorite: false
                    error: null
        '401':
          description: No autenticado. El token JWT está ausente o es inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                missing_token:
                  summary: Token ausente
                  value:
                    success: false
                    data: null
                    error:
                      message: "Unauthorized"
                invalid_token:
                  summary: Token inválido o expirado
                  value:
                    success: false
                    data: null
                    error:
                      message: "Unauthorized"
        '404':
          description: Libro no encontrado. No existe ningún libro con el `id` indicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                success: false
                data: null
                error:
                  message: "Book with id \"550e8400-e29b-41d4-a716-446655440000\" not found"
...
components:
  schemas:
...
    ApiError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "Validation failed"
        details:
          type: array
          items:
            type: string
          description: Detalles adicionales del error (opcional, usado en errores de validación múltiples)
          example: ["title is required", "authors is required"]
    ApiSuccessResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          enum: [true]
          description: Siempre true en respuestas exitosas
        data:
          type: object
          description: Datos de la respuesta
        error:
          type: 'null'
          description: Siempre null en respuestas exitosas
    ApiErrorResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          enum: [false]
          description: Siempre false en respuestas de error
        data:
          type: 'null'
          description: Siempre null en respuestas de error
        error:
          $ref: '#/components/schemas/ApiError'
...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Dirección de correo electrónico para enviar el enlace de restablecimiento
          example: "usuario@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          minLength: 1
          description: Token de restablecimiento recibido por correo electrónico
          example: "abc123def456ghi789..."
        newPassword:
          type: string
          minLength: 8
          description: Nueva contraseña (mínimo 8 caracteres)
          example: "nueva-contraseña-segura"

    # ============================================
    # HU-039: Favorites and Access Control Schemas
    # ============================================
    FavoriteToggleData:
      type: object
      required:
        - favorite
      properties:
        favorite:
          type: boolean
          description: |
            Estado del favorito tras la operación.
            `true` si el libro fue añadido a favoritos, `false` si fue eliminado.
          example: true
    FavoriteToggleResponse:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/FavoriteToggleData'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT de acceso almacenado en la cookie `access_token` (httpOnly).
        Se establece automáticamente al hacer login con `POST /auth/login`.
        Expira en 7 días. Puede renovarse con `POST /auth/refresh`.